text-backend: send key release to proper grab
When the input method grabs the keyboard, we must record the keys that
are pressed are send them to the default grab when they are released.

Without this, the client that had the default grab has no way of knowing
it missed key releases, since it does not receive a leave event.